+++ packages/ui/src/components/widgets/WidgetErrorWell.tsx
@@ -16,6 +16,7 @@ type Props = {
   relayerFeeProportion?: bigint | 0
   isHighRelayerServiceFee?: boolean
   isCapacityExceededError?: boolean
+  isCouldNotExecuteError?: boolean
   maxCapacity?: string
   supportsExternalLiquidity?: boolean
   containerCss?: Styles
@@ -29,6 +30,7 @@ export const WidgetErrorWell: FC<Props> = ({
   relayerFeeProportion,
   isHighRelayerServiceFee,
   isCapacityExceededError,
+  isCouldNotExecuteError,
   maxCapacity,
   supportsExternalLiquidity,
   containerCss
@@ -72,6 +74,31 @@ export const WidgetErrorWell: FC<Props> = ({
             Due to high demand, only {maxCapacity} {currency.symbol} can be
             bridged instantly. Set to max instant capacity or switch to the
             native route for unlimited capacity.
+          </Text>
+        </Flex>
+      )
+    } else if (
+      supportsExternalLiquidity &&
+      isCouldNotExecuteError &&
+      currency
+    ) {
+      return (
+        <Flex
+          align="center"
+          css={{
+            gap: '2',
+            p: '3',
+            backgroundColor: 'red2',
+            borderRadius: 12,
+            mb: '3',
+            ...containerCss
+          }}
+        >
+          <Box css={{ color: 'red10' }}>
+            <FontAwesomeIcon icon={faExclamationCircle} width={16} />
+          </Box>
+          <Text style="subtitle3" css={{ color: 'red12' }}>
+            {fetchQuoteErrorMessage} Please switch to the native route.
           </Text>
         </Flex>
       )